Moved output classes to output package
