+++ service/src/main/java/dk/dbc/opensearch/output/IndexHtmlBean.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package dk.dbc.opensearch.input;
+package dk.dbc.opensearch.output;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
